+++ src/api/api.ts
@@ -1,12 +1,12 @@
-import axios from "axios";
-import {getRegistrySetting} from "../utils/settings";
+import axios from "axios"
+import { getRegistrySetting } from "../utils/settings"
 
 const getPackage = async (packageName: string) => {
-  const registry = getRegistrySetting();
+  const registry = getRegistrySetting()
 
-  const result = await axios.get(`${registry}/${packageName}`);
+  const result = await axios.get(`${registry}/${packageName}`)
 
-  return result?.data;
-};
+  return result?.data
+}
 
-export {getPackage};
+export { getPackage }