+++ src/utils/messages.ts
@@ -1,22 +1,18 @@
-import * as vscode from "vscode";
-import {DO_NOT_SHOW_AGAIN, SUCCESS_MSG} from "../constants/generic";
-import Message from "../enums/Message";
-import {displayMessage} from "./helpers";
+import * as vscode from "vscode"
+import { DO_NOT_SHOW_AGAIN, SUCCESS_MSG } from "../constants/generic"
+import Message from "../enums/Message"
+import { displayMessage } from "./helpers"
 
 const getCompletedMessage = (context: vscode.ExtensionContext) => {
   if (!context.globalState.get(SUCCESS_MSG)) {
-    const completeMessageOptions = [DO_NOT_SHOW_AGAIN, "Ok"];
+    const completeMessageOptions = [DO_NOT_SHOW_AGAIN, "Ok"]
 
-    displayMessage(
-      "Version update complete.",
-      Message.INFO,
-      completeMessageOptions
-    ).then((value) => {
+    displayMessage("Version update complete.", Message.INFO, completeMessageOptions).then((value) => {
       if (value === DO_NOT_SHOW_AGAIN) {
-        context.globalState.update(SUCCESS_MSG, true);
+        context.globalState.update(SUCCESS_MSG, true)
       }
-    });
+    })
   }
-};
+}
 
-export {getCompletedMessage};
+export { getCompletedMessage }