+++ test/global.d.ts
@@ -1,0 +1,4 @@
+import { SuiteFunction, TestFunction } from "mocha"
+
+declare const suite: SuiteFunction
+declare const test: TestFunction